# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Badran Car Wash Supplies & Car Accessories Store API",
    "version": "1.0.0",
    "description": "Static OpenAPI snapshot for the Badran Store modular monolith API."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Search products",
        "parameters": [
          { "name": "categoryId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "brandId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "default": "productId" } },
          { "name": "sortDir", "in": "query", "schema": { "type": "string", "default": "asc" } }
        ],
        "responses": {
          "200": { "description": "Product page retrieved" }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Get product",
        "parameters": [{ "$ref": "#/components/parameters/ProductIdPath" }],
        "responses": {
          "200": { "description": "Product retrieved" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/api/v1/products/internal/{id}/deduct-stock": {
      "put": {
        "tags": ["Catalog"],
        "summary": "Deduct stock",
        "parameters": [
          { "$ref": "#/components/parameters/ProductIdPath" },
          { "name": "quantity", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 1 }, "example": 1 }
        ],
        "responses": {
          "200": { "description": "Stock deducted" },
          "400": { "description": "Invalid quantity or insufficient stock" },
          "403": { "description": "Admin role required" },
          "404": { "description": "Product not found" }
        }
      }
    },
...
    "/api/v1/wishlist": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Get wishlist",
        "parameters": [{ "$ref": "#/components/parameters/UserIdHeader" }],
        "responses": { "200": { "description": "Wishlist retrieved" } }
      }
    },
    "/api/v1/wishlist/{productId}": {
      "post": {
        "tags": ["Catalog"],
        "summary": "Add wishlist item",
        "parameters": [{ "$ref": "#/components/parameters/UserIdHeader" }, { "$ref": "#/components/parameters/ProductIdPath" }],
        "responses": { "200": { "description": "Product added to wishlist" }, "400": { "description": "Already in wishlist" }, "404": { "description": "Product not found" } }
      },
      "delete": {
        "tags": ["Catalog"],
        "summary": "Remove wishlist item",
        "parameters": [{ "$ref": "#/components/parameters/UserIdHeader" }, { "$ref": "#/components/parameters/ProductIdPath" }],
        "responses": { "200": { "description": "Product removed from wishlist" } }
      }
    },
    "/api/v1/reviews/product/{productId}": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Get product reviews",
        "parameters": [{ "$ref": "#/components/parameters/ProductIdPath" }],
        "responses": { "200": { "description": "Reviews retrieved" } }
      },
      "post": {
        "tags": ["Catalog"],
        "summary": "Add product review",
        "parameters": [
          { "$ref": "#/components/parameters/UserIdHeader" },
          { "$ref": "#/components/parameters/ProductIdPath" },
          { "name": "orderId", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 },
          { "name": "rating", "in": "query", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 5 }, "example": 5 },
          { "name": "comment", "in": "query", "schema": { "type": "string" }, "example": "Excellent" }
        ],
        "responses": { "200": { "description": "Review added" }, "400": { "description": "Invalid review request" }, "403": { "description": "Order does not belong to user" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "UserIdHeader": { "name": "X-User-Id", "in": "header", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 },
...
      "ProductIdPath": { "name": "productId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 },
...
    },
...
  }
}